+++ packages/sampler/.storybook/manager.js
@@ -1,6 +1,6 @@
 import ri from '@enact/ui/resolution';
-import {addons} from '@storybook/manager-api';
-import {create} from '@storybook/theming';
+import {addons} from 'storybook/manager-api';
+import {create} from 'storybook/theming';
 
 addons.setConfig({
 	theme: create({