+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `spotlight` methods `getPausedInstance` to get the name of the paused instance
+
+### Fixed
+
+- `spotlight` to not prioritize elements which are invisible due to overflow as next spottable elements
+
+## [5.1.0] - 2025-07-11
+
+No significant changes.
+
 ## [5.0.0] - 2025-06-13
 
 No significant changes.